# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cobros y Pagos API",
    "version": "0.1.0",
    "description": "Centavos enteros DOP. Sin impuestos ni amortización. Idempotency-Key requerido en operaciones financieras."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "summary": "Iniciar sesión",
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "maxLength": 200,
                    "format": "email",
                    "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  }
                },
                "required": [
                  "email",
                  "password"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operación completada"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Sesión requerida"
          },
          "403": {
            "description": "No autorizado"
          },
          "409": {
            "description": "Conflicto o límite"
          },
          "422": {
            "description": "Regla de negocio"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "summary": "Usuario actual",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación completada"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Sesión requerida"
          },
          "403": {
            "description": "No autorizado"
          },
          "409": {
            "description": "Conflicto o límite"
          },
          "422": {
            "description": "Regla de negocio"
          }
        }
      }
    },
    "/api/snapshot": {
      "get": {
        "summary": "Vista operacional autorizada",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación completada"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Sesión requerida"
          },
          "403": {
            "description": "No autorizado"
          },
          "409": {
            "description": "Conflicto o límite"
          },
          "422": {
            "description": "Regla de negocio"
          }
        }
      }
    },
    "/api/cargos": {
      "get": {
        "summary": "Consultar charges",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación completada"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Sesión requerida"
          },
          "403": {
            "description": "No autorizado"
          },
          "409": {
            "description": "Conflicto o límite"
          },
          "422": {
            "description": "Regla de negocio"
          }
        }
      },
      "post": {
        "summary": "Crear cargo",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 100
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "clientId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80
                  },
                  "service": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 160
                  },
                  "amount": {
                    "type": "integer",
                    "exclusiveMinimum": 0,
                    "maximum": 1000000000
                  },
                  "dueDate": {
                    "type": "string",
                    "format": "date",
                    "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
                  },
                  "required": {
                    "default": false,
                    "type": "boolean"
                  }
                },
                "required": [
                  "clientId",
                  "service",
                  "amount",
                  "dueDate",
                  "required"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operación completada"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Sesión requerida"
          },
          "403": {
            "description": "No autorizado"
          },
          "409": {
            "description": "Conflicto o límite"
          },
          "422": {
            "description": "Regla de negocio"
          }
        }
      }
    },
    "/api/descargos": {
      "get": {
        "summary": "Consultar payouts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación completada"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Sesión requerida"
          },
          "403": {
            "description": "No autorizado"
          },
          "409": {
            "description": "Conflicto o límite"
          },
          "422": {
            "description": "Regla de negocio"
          }
        }
      },
      "post": {
        "summary": "Autorizar pago a cliente",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 100
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "clientId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80
                  },
                  "collectorId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80
                  },
                  "concept": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 160
                  },
                  "amount": {
                    "type": "integer",
                    "exclusiveMinimum": 0,
                    "maximum": 1000000000
                  }
                },
                "required": [
                  "clientId",
                  "collectorId",
                  "concept",
                  "amount"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operación completada"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Sesión requerida"
          },
          "403": {
            "description": "No autorizado"
          },
          "409": {
            "description": "Conflicto o límite"
          },
          "422": {
            "description": "Regla de negocio"
          }
        }
      }
    },
    "/api/clientes": {
      "get": {
        "summary": "Consultar clients",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación completada"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Sesión requerida"
          },
          "403": {
            "description": "No autorizado"
          },
          "409": {
            "description": "Conflicto o límite"
          },
          "422": {
            "description": "Regla de negocio"
          }
        }
      }
    },
    "/api/rutas": {
      "get": {
        "summary": "Consultar routes",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación completada"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Sesión requerida"
          },
          "403": {
            "description": "No autorizado"
          },
          "409": {
            "description": "Conflicto o límite"
          },
          "422": {
            "description": "Regla de negocio"
          }
        }
      }
    },
    "/api/cuadres": {
      "get": {
        "summary": "Consultar settlements",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación completada"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Sesión requerida"
          },
          "403": {
            "description": "No autorizado"
          },
          "409": {
            "description": "Conflicto o límite"
          },
          "422": {
            "description": "Regla de negocio"
          }
        }
      },
      "post": {
        "summary": "Cerrar jornada con diferencia cero",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 100
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "collectorId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80
                  },
                  "date": {
                    "type": "string",
                    "format": "date",
                    "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
                  }
                },
                "required": [
                  "collectorId",
                  "date"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operación completada"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Sesión requerida"
          },
          "403": {
            "description": "No autorizado"
          },
          "409": {
            "description": "Conflicto o límite"
          },
          "422": {
            "description": "Regla de negocio"
          }
        }
      }
    },
...
    "/api/cuadres/preview": {
      "get": {
        "summary": "Calcular cuadre desde el libro de movimientos",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "collectorId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación completada"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Sesión requerida"
          },
          "403": {
            "description": "No autorizado"
          },
          "409": {
            "description": "Conflicto o límite"
          },
          "422": {
            "description": "Regla de negocio"
          }
        }
      }
    },
...
  }
}